+++ flutter_inappwebview_ios/ios/flutter_inappwebview_ios/Sources/flutter_inappwebview_ios/InAppBrowser/InAppBrowserNavigationController.swift
@@ -1,0 +1,16 @@
+//
+//  InAppBrowserNavigationController.swift
+//  flutter_inappwebview
+//
+//  Created by Lorenzo Pichilli on 14/02/21.
+//
+
+import Foundation
+import UIKit
+
+public class InAppBrowserNavigationController: UINavigationController {
+    deinit {
+        debugPrint("InAppBrowserNavigationController - dealloc")
+        UIApplication.shared.delegate?.window??.makeKeyAndVisible()
+    }
+}